Revert "Refactor for code reuseability in docker"
